Crate ap_start

Routines for booting up secondary CPU cores, aka application processors (APs) on x86_64.

Statics

AP_READY_FLAG

An atomic flag used for synchronizing progress between the BSP and the AP that is currently being booted. False means the AP hasn’t started or hasn’t yet finished booting.

Functions

insert_ap_stack

Insert a new stack that was allocated for the AP with the given apic_id.

kstart_ap

Entry to rust for an AP. The arguments must match the invocation order in “ap_boot.asm”